pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ildan-memforge"
version = "0.5.1"
description = "MemForge: portable, agent-neutral persistent memory for coding agents (ILDAN)."
readme = "README.md"
license = { text = "Apache-2.0" }
authors = [
    { name = "ILDAN" },
]
requires-python = ">=3.10"
dependencies = [
    "PyYAML>=6.0",
    "watchdog>=4.0",
]

[project.optional-dependencies]
dlp = [
    "detect-secrets>=1.5",
]
dev = [
    "pytest>=8",
    "pip-audit>=2.7",
]

[project.urls]
Homepage = "https://github.com/ildan-ai/memforge"
Issues = "https://github.com/ildan-ai/memforge/issues"

[project.scripts]
memforge = "memforge.cli._dispatch:main"
memory-audit = "memforge.cli.audit:main"
memory-audit-deep = "memforge.cli.audit_deep:main"
memory-promote = "memforge.cli.promote:main"
memory-audit-log = "memforge.cli.audit_log:main"
memory-cluster-suggest = "memforge.cli.cluster_suggest:main"
memory-dedup = "memforge.cli.dedup:main"
memory-dlp-scan = "memforge.cli.dlp_scan:main"
memory-frontmatter-backfill = "memforge.cli.frontmatter_backfill:main"
memory-index-gen = "memforge.cli.index_gen:main"
memory-link-rewriter = "memforge.cli.link_rewriter:main"
memory-preamble-extract = "memforge.cli.preamble_extract:main"
memory-query = "memforge.cli.query:main"
memory-rollup = "memforge.cli.rollup:main"
memory-watch = "memforge.cli.watch:main"
agents-md-gen = "memforge.cli.agents_md_gen:main"
memforge-resolve = "memforge.cli.resolve:main"
memforge-migrate-claim-block = "memforge.cli.migrate_claim_block:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
memforge = ["py.typed"]
